+++ game/src/main/java/net/demilich/metastone/game/spells/BuffSpell.java
@@ -8,20 +8,34 @@ import net.demilich.metastone.game.entities.EntityType;
 import net.demilich.metastone.game.entities.heroes.Hero;
 import net.demilich.metastone.game.entities.weapons.Weapon;
 import net.demilich.metastone.game.events.MaxHpIncreasedEvent;
+import net.demilich.metastone.game.spells.aura.ModifyBuffSpellAura;
 import net.demilich.metastone.game.spells.desc.SpellArg;
 import net.demilich.metastone.game.spells.desc.SpellDesc;
+import net.demilich.metastone.game.spells.desc.aura.AuraArg;
 import net.demilich.metastone.game.spells.desc.valueprovider.ValueProvider;
 import net.demilich.metastone.game.targeting.EntityReference;
 import net.demilich.metastone.game.cards.Attribute;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import java.util.List;
 import java.util.Map;
 
 /**
  * Gives the {@code target} a stats boost of either *+ {@link SpellArg#VALUE} / + {@link SpellArg#VALUE}) or (+ {@link
  * SpellArg#ATTACK_BONUS} / + {@link SpellArg#HP_BONUS} ). If the target is a {@link Hero}, like {@link
  * EntityReference#FRIENDLY_HERO}, {@link SpellArg#ARMOR_BONUS} will give the hero armor.
+ * <p>
+ * If a {@link net.demilich.metastone.game.spells.aura.ModifyBuffSpellAura} is in play and matches the {@code source},
+ * this effect will add the aura's specified attack and HP bonuses to the buff given by this spell <b>only</b> if the
+ * buff spell has the specified key.
+ * <ul>
+ * <li>If this spell has {@link SpellArg#VALUE}, both the aura's {@link AuraArg#ATTACK_BONUS}, {@link AuraArg#HP_BONUS}
+ * and {@link AuraArg#VALUE} values will be used.</li>
+ * <li>If this spell has {@link SpellArg#ATTACK_BONUS}, the aura's {@link AuraArg#ATTACK_BONUS} will be used.</li>
+ * <li>If this spell has {@link SpellArg#HP_BONUS}, the aura's {@link AuraArg#HP_BONUS} will be used</li>
+ * <li>{@link AuraArg#VALUE} is always added.</li>
+ * </ul>
  * <p>
  * For example, this trigger implements "Whenever you cast a spell, gain Armor equal to its Cost:"
  * <pre>
@@ -133,6 +147,7 @@ public class BuffSpell extends RevertableSpell {
 	protected void onCast(GameContext context, Player player, SpellDesc desc, Entity source, Entity target) {
 		super.onCast(context, player, desc, source, target);
 		checkArguments(LOGGER, context, source, desc, SpellArg.ATTACK_BONUS, SpellArg.HP_BONUS, SpellArg.ARMOR_BONUS, SpellArg.VALUE, SpellArg.REVERT_TRIGGER);
+
 		int attackBonus = desc.getValue(SpellArg.ATTACK_BONUS, context, player, target, source, 0);
 		int hpBonus = desc.getValue(SpellArg.HP_BONUS, context, player, target, source, 0);
 		int armorBonus = desc.getValue(SpellArg.ARMOR_BONUS, context, player, target, source, 0);
@@ -145,6 +160,29 @@ public class BuffSpell extends RevertableSpell {
 				attackBonus = hpBonus = value;
 			}
 		}
+
+		// Read aura bonuses
+		List<ModifyBuffSpellAura> auras = SpellUtils.getAuras(context, ModifyBuffSpellAura.class, source);
+
+		for (ModifyBuffSpellAura aura : auras) {
+			int auraAttackBonus = aura.getDesc().getValue(AuraArg.ATTACK_BONUS, context, player, target, source, 0);
+			int auraHpBonus = aura.getDesc().getValue(AuraArg.HP_BONUS, context, player, target, source, 0);
+			int auraValueBonus = aura.getDesc().getValue(AuraArg.VALUE, context, player, target, source, 0);
+			if (desc.containsKey(SpellArg.VALUE)) {
+				attackBonus += auraAttackBonus + auraValueBonus;
+				hpBonus += auraHpBonus + auraValueBonus;
+			} else {
+				if (desc.containsKey(SpellArg.ATTACK_BONUS)) {
+					attackBonus += auraAttackBonus + auraValueBonus;
+					hpBonus += auraValueBonus;
+				}
+				if (desc.containsKey(SpellArg.HP_BONUS)) {
+					attackBonus += auraValueBonus;
+					hpBonus += auraHpBonus + auraValueBonus;
+				}
+			}
+		}
+
 
 		if (attackBonus != 0) {
 			if (target instanceof Hero) {